% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{grpc_client}
\alias{grpc_client}
\title{Build a client handle}
\usage{
grpc_client(impl, channel)
}
\arguments{
\item{channel}{what to connect to}

\item{stubs}{the stub of the service}
}
\value{
client handle
}
\description{
Build a client handle
}
