% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grpc_debug}
\alias{grpc_debug}
\title{Check, enable or disable debug messages at the C++ level}
\usage{
grpc_debug(enable)
}
\arguments{
\item{enable}{booelan or missing}
}
\value{
boolean
}
\description{
Check, enable or disable debug messages at the C++ level
}
