% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grpc_default_hooks}
\alias{grpc_default_hooks}
\title{Provides a list of logging functions to be passed to gRPC server}
\usage{
grpc_default_hooks()
}
\value{
\code{list}
}
\description{
You may want to override this with custom functions, eg registering the gRPC service in Hashicorp's consul.io or similar after \code{server_start} etc.
}
