% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{helloworld}
\alias{helloworld}
\title{Example gRPC service}
\usage{
helloworld()
}
\description{
Reads a message with a name and returns a message greeting the name.
}
\references{
\url{https://github.com/grpc/grpc/tree/master/examples/cpp/helloworld}
}
