% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{newResponse}
\alias{newResponse}
\title{Construct a new ProtoBuf of ResponseType}
\usage{
newResponse(..., WFUN = sys.function(-1))
}
\arguments{
\item{...}{threaded through to the ProtoBuf constructor}

\item{WFUN}{A GRPC service method with RequestType and ResponseType attributes}
}
\value{

}
\description{
Construct a new ProtoBuf of ResponseType
}
