% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_services}
\alias{read_services}
\title{Create stub object from protobuf spec}
\usage{
read_services(file)
}
\arguments{
\item{file}{the spec file}
}
\value{
a stub data structure
}
\description{
Create stub object from protobuf spec
}
