% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{grpc_client}
\alias{grpc_client}
\title{Build a gRPC client stub}
\usage{
grpc_client(services, channel)
}
\arguments{
\item{services}{A list describing the service methods, typically from \code{read_services()}.
Each element corresponds to a gRPC method and should be a list containing:
\itemize{
  \item \code{name}: The fully qualified gRPC method string (e.g., "/package.Service/MethodName").
  \item \code{RequestTypeName}: Character string: the fully qualified name of the request message type.
  \item \code{ResponseTypeName}: Character string: the fully qualified name of the response message type.
}}

\item{channel}{A string specifying the server address and port (e.g., "localhost:50051").}
}
\value{
A list where names correspond to the simple service method names.
  Each element is a list with \code{build(...)} and \code{call(...)} functions.
}
\description{
Creates a list of functions that can be used to call methods on a gRPC service.
}
