% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{grpc_client}
\alias{grpc_client}
\title{Build a gRPC client stub}
\usage{
grpc_client(services, channel)
}
\arguments{
\item{services}{A list describing the service methods. Typically generated by
a function like \code{read_services()}. Each element of the list should
correspond to a gRPC method. The name of the list element should be the
simple method name (e.g., "SayHello"). Each element itself should be a list
containing:
\itemize{
  \item \code{name}: The fully qualified gRPC method string (e.g., "/package.Service/MethodName").
  \item \code{RequestType}: The \code{RProtoBuf::Descriptor} object for the request message type.
  \item \code{ResponseType}: The \code{RProtoBuf::Descriptor} object for the response message type.
}}

\item{channel}{A string specifying the server address and port (e.g., "localhost:50051").}
}
\value{
A list where names correspond to the simple service method names.
  Each element is itself a list with two functions:
  \itemize{
    \item \code{build(...)}: A function to construct a new request message object of the
          correct type for that method. Arguments are passed to \code{RProtoBuf::new()}.
    \item \code{call(request_message, metadata = list())}: A function to execute the RPC.
          It takes the RProtoBuf request message object and an optional named R list for metadata.
          It returns the deserialized RProtoBuf response message object.
  }
}
\description{
Creates a list of functions that can be used to call methods on a gRPC service.
}
