% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{start_server}
\alias{start_server}
\title{Start a gRPC server}
\usage{
start_server(impl, channel, hooks = list(prestart = function(params)
  cat("gRPC service started on port", params$port, "\\n")))
}
\arguments{
\item{impl}{an implementation of a proto service}

\item{channel}{a channel string in 'host:port' format}

\item{hooks}{list of R function(s) with \code{params} argument as a list to be run at different parts of the C++ calls. Supported hooks: \code{prebind}, \code{prestart}, \code{preinit}, \code{postinit}, \code{preshutdown}, \code{stopped}}
}
\value{
none
}
\description{
Start a gRPC server
}
