% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{start_server}
\alias{start_server}
\title{Start a gRPC server}
\usage{
start_server(impl, channel, on_start_hook = function(port)
  cat("gRPC service started on port", port, "\\n"))
}
\arguments{
\item{impl}{an implementation of a proto service}

\item{channel}{a channel string in 'host:port' format}

\item{on_start_hook}{R function with \code{port} argument to run when the gRPC service started}
}
\value{
none
}
\description{
Start a gRPC server
}
