% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{start_server}
\alias{start_server}
\title{Start a gRPC server}
\usage{
start_server(impl, hoststring)
}
\arguments{
\item{impl}{an implementation of a proto service}

\item{host}{a hoststring in 'host:port' format}
}
\value{
none
}
\description{
Start a gRPC server
}
